+++ src/commonMain/kotlin/events/Events.kt
@@ -5,6 +5,6 @@ import com.soywiz.korge.view.View
 
 data class BulletHitEvent(val target: View, val damage: Double)
 
-data class EnemyAttackEvent(val damage: Int)
+data class EnemyAttackEvent(val damage: Double)
 
 data class ChangeWeapon(val weapon: Weapon)
